refactor(channel): extract dispatch.rs from mod.rs
